requirements.txt
six
tensorboard
tensorflow
tensorflow-estimator
termcolor
typing-extensions
urllib3
Werkzeug
wrapt
zipp
click
colorama
dataclasses
Flask
gast
grpcio
h5py
idna
importlib-metadata
sklearn
scikit-learn==0.22
absl-py
astor
cached-property
cachetools
certifi
chardet
itsdangerous
Jinja2
joblib
Keras-Applications
Keras-Preprocessing
Markdown
MarkupSafe
pyasn1-modules
requests
requests-oauthlib
rsa
numpy
oauthlib
opt-einsum
pickle-mixin
Pillow
protobuf
pyasn1